requirements.txt
# Additional requirements go here
django-mptt==0.13.4
